+++ src/pages/index.tsx
@@ -1,0 +1,19 @@
+import type { NextPage } from 'next'
+import Head from 'next/head'
+import Image from 'next/image'
+import styles from '../../styles/Home.module.css'
+
+const Home: NextPage = () => {
+  return (
+    <div className={styles.container}>
+      <Head>
+        <title>Create Next App</title>
+        <meta name="description" content="Generated by create next app" />
+      </Head>
+    
+     
+    </div>
+  )
+}
+
+export default Home
