+++ styles/Home.module.css
@@ -19,6 +19,7 @@
   border-top: 1px solid #eaeaea;
   justify-content: center;
   align-items: center;
+  background: red;
 }
 
 .footer a {
